commit40

+++ demo/bill/views.py
@@ -491,7 +491,7 @@ def add_invoice(request):
             item_details_id=request.POST.get('item'+str(i))
             item_details=item.objects.get(id=item_details_id)
             quantity=int(request.POST.get('quantity'+str(i)))
-            rate=int(request.POST.get('rate'+str(i)))
+            rate=float(request.POST.get('rate'+str(i)))
             amount=quantity*rate
             taxable_amt+=amount
             avg_sgst+=item_details.sgst
